# Add delete implementation