--- a/.gitignore
+++ b/.gitignore
@@ -6,3 +6,5 @@
 *.mprof
 
 vendor/github.com/buger/goterm/
+prof.cpu
+prof.mem