**Description**:

This pull request adds a delete method to jsonparser, allowing users to remove json at a given path.

**Benchmark before change**:
```
BenchmarkJsonParserLarge-4                    	  100000	     73291 ns/op	     752 B/op	      94 allocs/op
BenchmarkJsonParserMedium-4                   	  500000	     12786 ns/op	      40 B/op	       5 allocs/op
BenchmarkJsonParserEachKeyManualMedium-4      	 1000000	      7033 ns/op	     120 B/op	       3 allocs/op
BenchmarkJsonParserEachKeyStructMedium-4      	 1000000	      7535 ns/op	     672 B/op	      12 allocs/op
BenchmarkJsonParserObjectEachStructMedium-4   	  500000	     11908 ns/op	     664 B/op	      16 allocs/op
BenchmarkJsonParserSmall-4                    	10000000	      1179 ns/op	      32 B/op	       4 allocs/op
BenchmarkJsonParserEachKeyManualSmall-4       	10000000	       733 ns/op	      80 B/op	       2 allocs/op
BenchmarkJsonParserEachKeyStructSmall-4       	10000000	      1035 ns/op	     256 B/op	       9 allocs/op
BenchmarkJsonParserObjectEachStructSmall-4    	10000000	       915 ns/op	     240 B/op	       8 allocs/op
BenchmarkJsonParserSetSmall-4                 	 5000000	      1511 ns/op	     848 B/op	       9 allocs/op
PASS
ok  	github.com/buger/jsonparser/benchmark	87.053s
```
**Benchmark after change**:

```
BenchmarkJsonParserLarge-4                    	  100000	     60300 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserMedium-4                   	 1000000	     10371 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserDelMedium-4                	 1000000	      8493 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyManualMedium-4      	 1000000	      6847 ns/op	     112 B/op	       2 allocs/op
BenchmarkJsonParserEachKeyStructMedium-4      	 1000000	      7224 ns/op	     672 B/op	      12 allocs/op
BenchmarkJsonParserObjectEachStructMedium-4   	 1000000	     10331 ns/op	     624 B/op	      11 allocs/op
BenchmarkJsonParserSmall-4                    	10000000	       910 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyManualSmall-4       	10000000	       717 ns/op	      80 B/op	       2 allocs/op
BenchmarkJsonParserEachKeyStructSmall-4       	10000000	      1012 ns/op	     256 B/op	       9 allocs/op
BenchmarkJsonParserObjectEachStructSmall-4    	10000000	       910 ns/op	     240 B/op	       8 allocs/op
BenchmarkJsonParserSetSmall-4                 	 5000000	      1242 ns/op	     816 B/op	       5 allocs/op
BenchmarkJsonParserDelSmall-4                 	 5000000	      1511 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/buger/jsonparser/benchmark	106.127s
```