@buger 2017-07-03
Nice one! 

I think better rename it to `Delete` to follow standard naming. Also will be great if you update docus as well. 

Thank you!