@pytlesk4 2017-07-05
Thanks, just updated the docs and changed the name to Delete!